Move workload stats into PartitionHandler and out of the individual storage handlers
